chore: moved carousel to carousel screen
